refactor filenames and utils
