+++ tests/e2e/test_koperator_suite.go
@@ -50,3 +50,12 @@ var _ = BeforeSuite(func() {
 		Expect(len(pods)).To(Not(BeZero()))
 	})
 })
+
+var _ = When("Testing e2e test altogether", Ordered, func() {
+	//testInstall()
+	//testProduceConsumeInternal()
+	testProduceConsumeExternal()
+	// testUninstallZookeeperCluster()
+	// testUninstallKafkaCluster()
+	//	testUninstall()
+})
