+++ tests/e2e/test_uninstall_cluster.go
