+++ tests/e2e/utils.go
@@ -15,13 +15,7 @@
 package e2e
 
 import (
-	"fmt"
 	"sort"
-	"strings"
-)
-
-const (
-	kubectlNotFoundErrorMsg = "NotFound"
 )
 
 // stringSlicesUnion returns the union of the slices from argument.
@@ -48,39 +42,4 @@ func stringSlicesUnion(sliceA, sliceB []string) []string {
 		}
 	}
 	return union
-}
-
-// kubectlArgExtender extends the kubectl arguments and log message based on the parameters
-func kubectlArgExtender(args []string, logMsg, selector, names, namespace string, extraArgs []string) (string, []string) {
-	if selector != "" {
-		logMsg = fmt.Sprintf("%s selector: '%s'", logMsg, selector)
-		args = append(args, fmt.Sprintf("--selector=%s", selector))
-	} else if names != "" {
-		logMsg = fmt.Sprintf("%s name(s): '%s'", logMsg, names)
-		args = append(args, names)
-	}
-	if namespace != "" {
-		logMsg = fmt.Sprintf("%s namespace: '%s'", logMsg, namespace)
-	}
-	if len(extraArgs) != 0 {
-		logMsg = fmt.Sprintf("%s extraArgs: '%s'", logMsg, extraArgs)
-		args = append(args, extraArgs...)
-	}
-	return logMsg, args
-}
-
-// kubectlRemoveWarning removes those elements from the outputSlice parameter which contains kubectl warning message.
-func kubectlRemoveWarnings(outputSlice []string) []string {
-	// Remove warning message pollution from the output
-	result := make([]string, 0, len(outputSlice))
-	for i := range outputSlice {
-		if !strings.Contains(outputSlice[i], "Warning:") {
-			result = append(result, outputSlice[i])
-		}
-	}
-	return result
-}
-
-func isKubectlNotFoundError(err error) bool {
-	return err != nil && strings.Contains(err.Error(), kubectlNotFoundErrorMsg)
 }
